adds uniform_zoom to ImageDataGenerator and random_zoom
This modification gives the user the ability to define, if the zoom transformation should be the same on the xy axes or seperate for each axis. The parameter has to be introduced in the main keras library as well.

Hope this is the right way to suggest such change. Sorry if not, but I never did this before :)