Remove unnecessary cast to string
In the documentation it is set that the column should contain
filenames of filepaths, which are strings. Actually casting
to string will confuse the user as if the user's dataframe contains
integers for example here it will work but down his/her ml/data pipeline
could fail, even though it should have catched earlier.