Standarize arguments in function/method calls and instantiation (#121)
### Summary
At the moment there is no coding standard when the parameters of a method or instantiating a class are many. This PR standardizes them both. This allows better readability of the code, for example the instantiation of `DataframeIterator`, `DirectoryIterator` and `NumpyArrayIterator` in their respective `flow_from_` methods of `ImageDataGenerator` becomes clearer and easier to track.

### PR Overview

- [ n] This PR requires new unit tests [y/n] (make sure tests are included)
- [n ] This PR requires to update the documentation [y/n] (make sure the docs are up-to-date)
- [y ] This PR is backwards compatible [y/n]
- [n ] This PR changes the current API [y/n] (all API changes need to be approved by fchollet)