Default directory (#124)
### Summary
The documentation in `flow_from_dataframe` and `DataframeIterator` states that if `directory` is None then the filenames are taken as absolute paths. I agree with this behavior, and I think this should be the default behavior if no directory parameter is given. This PR sets the default directory parameter as None and adds the corresponding tests. With this PR is enough to have the absolute paths in a Dataframe and just send that dataframe as parameter to `flow_from_dataframe`

Closses #67 #56 #116 

### PR Overview

- [y] This PR requires new unit tests [y/n] (make sure tests are included)
- [n] This PR requires to update the documentation [y/n] (make sure the docs are up-to-date)
- [y] This PR is backwards compatible [y/n]
- [n] This PR changes the current API [y/n] (all API changes need to be approved by fchollet)