Update CI (#159)
### Summary
Update CI, basically separate every test component.
- PEP8 (no dependencies installed at all)
- PyTests (install test dependencies), runs for Python 2.6, 3.6 and Keras PyPI and Keras HEAD
- Imports (install package dependencies), runs for Python 2.6, 3.6. Basically imports the package and modules as a check that nothing breaks.

I also change where some of the logic happens. Before there was some logic in `travis.yml` that in my opinion `shuoldn't` be there but in `setup.py`, in this way the user can actually execute the same that travis is doing by `pip install -e .[tests] && pytest tests` or  `pip install -e .[pep8] && flake8`. So much easier to debug without the need to push and create a pull request to reproduce travis execution (up to some complexity).

This should lead to a more robust check and specially avoid the error given during the release of 1.0.6. That error for example will be catch in two places now: 
- PEP8: as an unused import
- Imports: since pandas is not installed on that check.

NOTE: The phase caching the bug in the 1.0.6 should not be the Pytest phase, that actually should work. Since pandas is needed for unit testing.

By running these tests on the current state of the repo I already found several bugs in the code, several overhead variables being assigned but unused (specially in tests) and like 30 plus PEP8 violations.

### Related Issues
#154 

### PR Overview

- [ n] This PR requires new unit tests [y/n] (make sure tests are included)
- [ n] This PR requires to update the documentation [y/n] (make sure the docs are up-to-date)
- [ y] This PR is backwards compatible [y/n]
- [ n] This PR changes the current API [y/n] (all API changes need to be approved by fchollet)