Add validate filenames parameter to DataFrameIterator (#177)
### Summary
This PR allows the user to skip the validation of images filepaths in the pandas DataFrame to speed up the instantiation process. Slow instantiation times have been reported (#171) for a big amount of filenames.  
In addition this PR adds a warning to the user if validation is on and non-valid image filenames are found.
By default image filenames are validated.

### Related Issues
#171 

### PR Overview

- [ y] This PR requires new unit tests [y/n] (make sure tests are included)
- [ y] This PR requires to update the documentation [y/n] (make sure the docs are up-to-date)
- [ y] This PR is backwards compatible [y/n]
- [ n] This PR changes the current API [y/n] (all API changes need to be approved by fchollet)